+++ src/mantle/testing/class-installation-manager.php
@@ -219,17 +219,48 @@ class Installation_Manager {
 	/**
 	 * Define the site/home URLs to be set after the installation is loaded.
 	 *
+	 * @throws \InvalidArgumentException If the home or site URL is invalid.
+	 *
 	 * @param string|null $home Home URL.
 	 * @param string|null $site Site URL.
+	 * @param bool        $set_tests_domain Whether to set WP_TESTS_DOMAIN constant to match the home URL.
 	 */
-	public function with_url( ?string $home = null, ?string $site = null ): static {
+	public function with_url( ?string $home = null, ?string $site = null, bool $set_tests_domain = true ): static {
 		if ( $home ) {
+			if ( ! filter_var( $home, FILTER_VALIDATE_URL ) ) {
+				throw new \InvalidArgumentException( 'Invalid home URL.' );
+			}
+
 			$this->with_option( 'home', $home );
+
+			if ( $set_tests_domain ) {
+				$this->before(
+					fn () => defined( 'WP_TESTS_DOMAIN' ) || define( 'WP_TESTS_DOMAIN', parse_url( $home, PHP_URL_HOST ) ), // phpcs:ignore WordPress.WP.AlternativeFunctions.parse_url_parse_url
+				);
+			}
 		}
 
 		if ( $site ) {
+			if ( ! filter_var( $site, FILTER_VALIDATE_URL ) ) {
+				throw new \InvalidArgumentException( 'Invalid site URL.' );
+			}
+
 			$this->with_option( 'siteurl', $site );
+
+			// Setup the default HTTP_HOST and HTTPS to make sure the site is installed properly.
+			$this->before( function () use ( $site ): void {
+				$_SERVER['HTTP_HOST'] = $_SERVER['SERVER_NAME'] = parse_url( $site, PHP_URL_HOST ); // phpcs:ignore WordPress.WP.AlternativeFunctions.parse_url_parse_url
+
+				if ( 'https' === parse_url( $site, PHP_URL_SCHEME ) ) { // phpcs:ignore WordPress.WP.AlternativeFunctions.parse_url_parse_url
+					$_SERVER['HTTPS'] = 'on';
+
+					defined( 'WP_TESTS_USE_HTTPS' ) || define( 'WP_TESTS_USE_HTTPS', true ); // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedConstantFound
+				} else {
+					unset( $_SERVER['HTTPS'] );
+				}
+			} );
 		}
+
 
 		return $this;
 	}