+++ src/mantle/testing/wordpress-bootstrap.php
@@ -160,6 +160,8 @@ if ( ! $installing_wp && '1' !== getenv( 'WP_TESTS_SKIP_INSTALL' ) ) {
 			WP_PHP_BINARY,
 			escapeshellarg( __DIR__ . '/install-wordpress.php' ),
 			$multisite,
+			WP_TESTS_DOMAIN,
+			! empty( $_SERVER['HTTPS'] ) ? '1' : '0',
 		],
 		$retval,
 	);