+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.4.4 - 2025-01-31
+
+### Fixed
+
+- Ensure that when setting the home URL for testing with `Installation_Manager::with_url()` also sets the test domain.
+- When setting a HTTPS home URL for testing, ensure that the site is installed using HTTPS.
+
 ## v1.4.3 - 2025-01-31
 
 ### Fixed